Refactor explore and dashboard screens
- Replaced the content of the explore and dashboard tabs with dedicated screen components.
- Moved the ExploreScreen and DashboardScreen implementations to their respective files in the screens directory.
- Updated imports and removed unused components to streamline the codebase.
